Split out schemas removal logic when removing operation groups
